add ColorPicker component
